% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{Fizz_Buzz_Game}
\alias{Fizz_Buzz_Game}
\title{FizzBuzz Game}
\usage{
Fizz_Buzz_Game(n)
}
\arguments{
\item{n}{A vector of integers.}
}
\value{
A vector of Fizz Buzz responses.
}
\description{
This function takes a vector of integers and returns the associated FizzBuzz responses.
}
\examples{
Fizz_Buzz_Game(1:15)

}
